refactor(snowman): extract pose updates (R3.9)
